+++ hyperkitty/tests/views/test_thread.py
@@ -25,13 +25,13 @@ import json
 import re
 import urllib
 from email.message import EmailMessage
+from unittest.mock import patch
 
 from django.contrib.auth.models import User
 from django.test import override_settings
 
 from bs4 import BeautifulSoup
 from django_mailman3.tests.utils import get_flash_messages
-from mock import patch
 
 from hyperkitty.lib.incoming import add_to_list
 from hyperkitty.models import Email, MailingList, Tag, Tagging, Thread